Made an important check as to whether or not response[this.name] actually exists, otherwise just set data=response
Basically for nested models, the root model given by Cake needs to be parsed by AppModel because Backbone doesn't like the format:

{
	Root: {},
	Secondary: {},
	Secondary: {}
}

but instead wants all the Root model's attributes at the lowest level. For the related models, however, the format is already compatible with Backbone but they were still being parsed by the AppModel. Thus we need to perform this check, which will in reality switch state between Root and Related models