Changed how add/reset are handled.
Because the add event is triggered every time a model is added, collectionHasAdded actually had the wrong signature. Therefore it was necessary to have two events to handle adding, one for each model, and one at the end. Thus fetch() had to be modified, as did the render function (because we couldn't hideLoading after only adding one model) and the reset and hideLoading functions too all to take this into account. The good news is that loadMore works beautifully.